+++ src/lib.rs
@@ -104,7 +104,7 @@ pub use crate::media::{
     VideoStream,
 };
 pub use crate::myplex::{
-    ConnectOptions, MyPlexClient, MyPlexDevice, MyPlexResource, ResourceConnection,
+    ConnectOptions, MyPlexClient, MyPlexDevice, MyPlexResource, ResourceConnection, UserState,
     WatchlistFilter, WatchlistItem, WatchlistKind, WatchlistOptions,
 };
 pub use crate::pagination::PageRange;