+++ src/myplex/mod.rs
@@ -23,11 +23,13 @@
 //! client at a wiremock replica.
 
 pub mod devices;
+pub mod metadata_provider;
 pub mod resources;
 pub mod watchlist;
 pub mod webhooks;
 
 pub use devices::MyPlexDevice;
+pub use metadata_provider::UserState;
 pub use resources::{ConnectOptions, MyPlexResource, ResourceConnection};
 pub use watchlist::{WatchlistFilter, WatchlistItem, WatchlistKind, WatchlistOptions};